board/build: refactoring
* generate all source defines and constants in headers
* concat everything generated in Target.h header
* make all detail::map functions constexpr to avoid calls in ISRs
* redesign io control for native inputs and outputs:
  * instead of reading pins one by one, read the entire ports
  * instead of toggling pins one by one, toggle entire port at once
* reduce the number of input read history to 16 from 32
* move st-specific map functions to new, internal header for st only
* updated core module and with it, io macros
* move arch/vendor specific config to yml files

+++ src/board/Board.h
@@ -265,12 +265,12 @@ namespace Board
 
         /// Structure containing digital input readings for a given input.
         /// Count represents total amount of readings stored in readings variable.
-        /// Readings variable contains up to last 32 readings where LSB bit is the
+        /// Readings variable contains up to 16 last readings where LSB bit is the
         /// newest reading, and MSB bit is the last.
         typedef struct
         {
             uint8_t  count;
-            uint32_t readings;
+            uint16_t readings;
         } dInReadings_t;
 
         enum class dataSource_t : uint8_t
